Improve e2e tests.
