--- project/test_framework/app/selectors.py
@@ -1,53 +1,0 @@
-civi_wiki_url = "http://civiwiki-test.herokuapp.com/"
-username = "unit1test"
-password = "unit1test"
-password_wrong = "unit123test"
-
-# Accessibility
-login_signup_button = ".login-btn"
-username_field = "#username"
-password_field = "#password"
-login_button = ".login-button"
-user_label = ".label-username"
-civi_wiki_logo = ".wordmark"
-nav_menu = ".nav-menu"
-logout_button = "a.dropdown-item:nth-child(5)"
-settings_button = "a.dropdown-item:nth-child(4)"
-my_profile_button = "a.dropdown-item:nth-child(2)"
-my_feed_button = "a.dropdown-item:nth-child(1)"
-
-# Settings
-email_field = "#email"
-first_name_field = "#first_name"
-last_name_field = "#last_name"
-location_field = ".location-info"
-location_set = "#autocomplete"
-activity = ".profile-civi-list"
-edit_profile_butoon = "a.btn"
-about_me_field = "#about_me"
-my_feed_list = "#feed-list"
-manage_categories = "#manage-categories"
-choose_button = "#choose-category"
-trending_issues = ".list-title"
-
-# Threads
-add_civi_button = ".add-civi"
-add_problem_button = "button.btn-small:nth-child(1)"
-add_cause_button = "button.btn-small:nth-child(2)"
-add_solution_button = "button.btn:nth-child(3)"
-civi_title_field = "#civi-title"
-civi_description_field = "#civi-body"
-civi_links_field = ".ms-sel-ctn > input:nth-child(1)"
-create_civi_button = ".create-new-civi"
-first_civi = ".civi-title > span:nth-child(1)"
-create_thread_button = ".new-thread"
-thread_title = "#thread-title"
-thread_summary = "#thread-body"
-thread_type_dropdown = ".new-thread-level-selection > div:nth-child(1) > div:nth-child(1)"
-thread_category_dropdown = ".new-thread-category-selection > div:nth-child(1) > div:nth-child(1) > input:nth-child(2)"
-thread_create_draft = ".draft-new-thread"
-thread_publish = "#js-publish-btn"
-add_response_button = "#add-new-response"
-response_title = "#response-title"
-response_body = "#response-body"
-response_confirm = ".create-new-response"
