+++ project/test_framework/my_feed/features/__init__.py
@@ -1,0 +1,2 @@
+from test_framework.accessability.features import *
+from test_framework.accessability.features.accessability import *
