--- project/test_framework/requirements.txt
@@ -1,2 +1,0 @@
-aloe==0.1.15
-selenium==3.8.0
