+++ requirements.txt
@@ -22,3 +22,7 @@ sunlight==1.2.9
 sunlight==1.2.9
 urllib3==1.22
 whitenoise==3.3.1
+
+# Testing
+aloe==0.1.18
+selenium==3.8.0
